refac(hellodex): reorganized components
