+++ src/pages/Hellodex/BattleRecord/BattleRecord.module.scss
