+++ src/pages/Hellodex/BattleRecord/BattleRecord.tsx
