+++ src/pages/Hellodex/BattleRecord/index.ts
@@ -1,0 +1,1 @@
+export * from './BattleRecord';
