+++ src/pages/Hellodex/GradientButton/GradientButton.module.scss
@@ -1,0 +1,64 @@
+@use 'config/colors';
+@use 'config/fonts';
+@use 'functions/color';
+@use 'mixins/font';
+@use 'mixins/position';
+@use 'mixins/spacing';
+@use 'mixins/transition';
+
+.container {
+  position: relative;
+  @include spacing.padding-xy($x: 35px, $y: 12px);
+  @include font.apply(fonts.$primary, ( font-weight: 300 ));
+  font-size: 14px;
+  color: colors.$transparent;
+  letter-spacing: 3px;
+  text-align: center;
+  text-transform: uppercase;
+  background: linear-gradient(75deg, #4A9AC7 5%, #F148CC 95%) 0 0 / 200% 100%;
+  background-clip: text;
+  border-radius: 100px;
+  @include transition.apply(color, box-shadow);
+
+  &.light {
+    box-shadow: (
+      0 0 1px color.alpha(colors.$black, 0.3),
+      0 0 5px color.alpha(colors.$gray-dark, 0.3),
+    );
+  }
+
+  &.dark {
+    box-shadow: (
+      0 0 1px color.alpha(colors.$white, 0.5),
+      0 0 10px color.alpha(colors.$black, 0.5),
+    );
+  }
+
+  &::before {
+    content: '';
+    @include position.absolute-fill;
+    background: linear-gradient(75deg, #4A9AC7 5%, #F148CC 95%) 0 0 / 200% 100%;
+    border-radius: 100px;
+    opacity: 0;
+    @include transition.apply((background-position, 1.5s), opacity);
+    z-index: -1;
+  }
+
+  &:hover {
+    color: colors.$white;
+    background-clip: border-box;
+
+    &::before {
+      background-position: 100% 0;
+      opacity: 0.65;
+    }
+  }
+
+  span, strong {
+    display: inline-block;
+  }
+
+  strong {
+    font-weight: 600;
+  }
+}
