+++ src/pages/Hellodex/GradientButton/index.ts
@@ -1,0 +1,1 @@
+export * from './GradientButton';
