+++ src/pages/Hellodex/InstanceButton/index.ts
@@ -1,0 +1,1 @@
+export * from './InstanceButton';
