+++ src/pages/Hellodex/PatronagePane/index.ts
@@ -1,0 +1,1 @@
+export * from './PatronagePane';
