+++ src/pages/Hellodex/SettingsPane/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingsPane';
